fix all register all bugs, new env.js ,fixing cloudinary,db

+++ src/utils/cloudinary.js
@@ -1,12 +1,21 @@
+import "../config/env.js"
+
+
+
 import { v2 as cloudinary } from "cloudinary";
-import fr from "fs";
+import fs from "fs";
 
 
 cloudinary.config({ 
-  cloud_name: process.env.CLOUDINARY_CLOUDE_NAME, 
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
   api_key: process.env.CLOUDINARY_API_KEY, 
-  api_secret: process.env.CLOUDINARY_API_SCREAT
+  api_secret: process.env.CLOUDINARY_API_SECRET
 });
+
+
+
+
+
 
 const uploadOnCloudinary = async (localFilePath) =>{
     try {
@@ -19,11 +28,23 @@ const uploadOnCloudinary = async (localFilePath) =>{
         console.log("done sir all set ", response.url);
         return response;
         
-    } catch(error) {
-        fs.unlinkSync(localFilePath)
-         // remove the temporary saved file as the upload operation got failed 
-        return null;
+     }
+    //  catch(error) {
+    //     fs.unlinkSync(localFilePath)
+    //      // remove the temporary saved file as the upload operation got failed 
+    //     return null;
+    // }
+    
+
+    catch (error) {
+    console.log("Cloudinary ERROR:", error); // 👈 ADD THIS
+
+    if (fs.existsSync(localFilePath)) {
+        fs.unlinkSync(localFilePath);
     }
+
+    return null;
+}
 }
